+++ performance/components/erb_cached_benchmark_component.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Performance
+  class ErbCachedBenchmarkComponent < HandlerBenchmarkComponent
+    include ViewComponent::ExperimentallyCacheable
+
+    cache do
+      [name]
+    end
+  end
+end