+++ performance/components/haml_uncached_benchmark_component.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module Performance
+  class HamlUncachedBenchmarkComponent < HandlerBenchmarkComponent
+  end
+end