+++ test/sandbox/app/components/cache_condition_component.rb
@@ -1,0 +1,22 @@
+# frozen_string_literal: true
+
+class CacheConditionComponent < ViewComponent::Base
+  include ViewComponent::ExperimentallyCacheable
+
+  cache_if :cache_enabled?
+  cache do
+    [foo]
+  end
+
+  attr_reader :foo
+
+  def initialize(foo:)
+    @foo = foo
+  end
+
+  private
+
+  def cache_enabled?
+    false
+  end
+end