+++ test/sandbox/app/components/cache_digestor_nested_partial_parent_component.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+class CacheDigestorNestedPartialParentComponent < ViewComponent::Base
+  include ViewComponent::ExperimentallyCacheable
+
+  cache do
+    [foo]
+  end
+
+  attr_reader :foo
+
+  def initialize(foo:)
+    @foo = foo
+  end
+end