+++ test/sandbox/app/components/cache_html_safety_component.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class CacheHtmlSafetyComponent < ViewComponent::Base
+  include ViewComponent::ExperimentallyCacheable
+
+  cache { [:h] }
+end