+++ test/sandbox/app/components/cache_indirect_parent_component.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class CacheIndirectParentComponent < ViewComponent::Base
+  include ViewComponent::ExperimentallyCacheable
+
+  cache { [:indirect] }
+
+  def indirect_child
+    render(CacheDigestorChildComponent.new)
+  end
+end