+++ test/sandbox/app/components/cache_self_referential_component.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+class CacheSelfReferentialComponent < ViewComponent::Base
+  include ViewComponent::ExperimentallyCacheable
+
+  cache { [:self] }
+
+  private
+
+  def render_self?
+    false
+  end
+end